added working cats logic
