+++ package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "@baltimorecounty/dotgov-components": "1.0.26",
     "@baltimorecounty/javascript-utilities": "^1.0.7",
-    "@baltimorecounty/react-filter-list": "^1.0.33",
+    "@baltimorecounty/react-filter-list": "^1.0.34",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.1.2",
