+++ public/index.html
@@ -32,7 +32,7 @@
         informationAbout:
           "If you're interested in adopting an animal, please have its Animal ID: {animalId} handy. Before calling or visiting our shelter, please review our page <a href='https://beta.baltimorecountymd.gov/departments/animalservices/how-to-adopt.html'>how to adopt</a>.",
         petStatus: "Adoptable",
-        workingCat: "false",
+        workingCats: "true",
       };
     </script>
     <!--********************************
