+++ src/App.js
@@ -2,7 +2,7 @@ import "@baltimorecounty/dotgov-components/lib/styles/dotgov.min.css";
 import "./App.css";
 
 import { Route, HashRouter as Router } from "react-router-dom";
-
+import { filters, filtersWorkingCats } from "./components/Filters";
 import { Config } from "@baltimorecounty/javascript-utilities";
 import Pet from "./pages/Pet";
 import PetsList from "./components/PetsList";
@@ -11,7 +11,7 @@ import React from "react";
 const { setConfig } = Config;
 
 const initialPetsEndpoint = "api/hub/pets/pets";
-
+const { workingCats } = window.pets;
 const testApiRoot = `https://testservices.baltimorecountymd.gov/${initialPetsEndpoint}`;
 const prodApiRoot = `https://services.baltimorecountymd.gov/${initialPetsEndpoint}`;
 
@@ -39,51 +39,6 @@ const configValues = {
 
 setConfig(configValues);
 
-const filters = [
-  {
-    targetApiField: "status",
-    value:
-      window.pets.petStatus ||
-      console.error("You must provide a pets.petStatus."),
-  },
-  // {
-  //   targetApiField: "workingcat",
-  //   value:
-  //     window.pets.workingCat,
-  // },
-  {
-    targetApiField: "recordsPerPage",
-    value: 10,
-  },
-  {
-    targetApiField: "petType",
-    displayName: "Species",
-    options: [
-      { value: "dog", label: "Dog" },
-      { value: "cat", label: "Cat" },
-      { value: "other", label: "Other" },
-    ],
-  },
-  {
-    targetApiField: "gender",
-    displayName: "Gender",
-    options: [
-      { value: "male", label: "Male" },
-      { value: "female", label: "Female" },
-      { value: "unknown", label: "Unknown" },
-    ],
-  },
-  {
-    targetApiField: "weight",
-    displayName: "Size",
-    options: [
-      { value: "small", label: "Small" },
-      { value: "medium", label: "Medium" },
-      { value: "large", label: "Large" },
-    ],
-  },
-];
-
 function App() {
   return (
     <React.Fragment>
@@ -91,7 +46,12 @@ function App() {
         <Route
           exact
           path="/"
-          component={(props) => <PetsList filters={filters} {...props} />}
+          component={(props) => (
+            <PetsList
+              filters={workingCats ? filtersWorkingCats : filters}
+              {...props}
+            />
+          )}
         />
         <Route exact path="/pets/:animalId" component={Pet} />
       </Router>
